requirements.txt
opencv-python>=4.8.0
face-recognition>=1.3.0
numpy>=1.21.0
firebase-admin>=6.0.0
pandas>=2.0.0
openpyxl>=3.1.0
Pillow>=10.0.0
